Move Project class to separate file
